Add caching support
Add support for the following config options `cacheable`, `dependenciesRoot`, and `parseComments`.

The default behaviour for the loader is to parse each `*.erb` file for JavaScript block comments starting with `uh-erb-loader-dependency`. If found, the script will be set to "cacheable", and will only be recompiled if one of these dependency files has been changed.